move `store/historyStore/model/queries.js` to `stores/services/history.services.js` and improve queries
